Move Code (Optimize Readablity)
