+++ GameLauncher/App/Classes/LauncherCore/Visuals/ProgressBar.cs
